pyproject.toml
[tool.poetry]
name = "certification-django"
version = "0.1.0"
description = ""
authors = ["Zidolch <yury.batin@gmail.com>"]
readme = "README.md"
packages = [{include = "certification_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
